requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
cssselect==1.2.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
jmespath==1.0.1
kaitaistruct==0.10
lxml==4.9.3
multidict==6.0.4
numpy==1.25.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
pyasn1==0.5.0
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.11.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
undetected-chromedriver==3.5.2
urllib3==2.0.3
w3lib==2.1.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zstandard==0.21.0
